backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.0.4
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
emoji==1.7.0
joblib==1.1.0
nltk==3.7
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
regex==2022.3.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.63.0
tzdata==2021.5
